HMRC-1438: Code review
