+++ config/routes.rb
@@ -160,7 +160,7 @@ Rails.application.routes.draw do
     resources :results, param: :category, only: %i[create], path: 'result'
 
     get 'faq', to: 'faq#index'
-    resources :faq_feedback, only: %i[index create], path: 'faq/feedback'
+    resources :faq_feedback, only: %i[index create], path: 'faq/feedback', controller: '/faq_feedback'
   end
 
   match '/search', as: :perform_search, via: %i[get post], to: 'search#search'
